# docs/openapi-doc.yml
openapi: 3.1.0
info:
    title: Firepit API
    version: 1.7.0
    description: |
        Firepit is a realtime communication platform built with Next.js 16, React 19, Bun, and Appwrite.

        This specification covers the public and first-party client API surface currently used by the application.
        Internal debug and test-only endpoints are intentionally omitted.
    contact:
        name: Firepit
        url: https://github.com/acarlson33/firepit
    license:
        name: GNU General Public License v3.0
        url: https://opensource.org/licenses/GPL-3.0
servers:
    - url: http://localhost:3000
      description: Local development server
    - url: https://firepit.example.com
      description: Example production server
...
components:
    securitySchemes:
        sessionCookie:
            type: apiKey
            in: cookie
            name: a_session_*
            description: Appwrite session cookie used by authenticated first-party clients
    schemas:
...
        ServerMember:
            type: object
            properties:
                userId:
                    type: string
                displayName:
                    type: string
                userName:
                    type: string
                avatarUrl:
                    type: string
                    format: uri
                joinedAt:
                    type: string
                    format: date-time
                roles:
                    type: array
                    items:
                        type: string
...
paths:
...
    /api/servers/{serverId}/members:
        get:
            tags: [Servers, Moderation]
            summary: List server members
            security:
                - sessionCookie: []
            parameters:
                - name: serverId
                  in: path
                  required: true
                  schema:
                      type: string
                - name: limit
                  in: query
                  schema:
                      type: integer
                - name: cursor
                  in: query
                  schema:
                      type: string
            responses:
                "200":
                    description: Server members
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    members:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/ServerMember"
                                    nextCursor:
                                        type: string
...
    /api/profiles/batch:
        post:
            tags: [Users]
            summary: Resolve multiple profiles in a single request
            security:
                - sessionCookie: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [userIds]
                            properties:
                                userIds:
                                    type: array
                                    items:
                                        type: string
            responses:
                "200":
                    description: Batch profile result
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: true
...